backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
firebase-admin
pydantic
